Rename qwde.pystock to dataprovider
