+++ .gitmodules
@@ -1,3 +1,3 @@
 [submodule "pystock-data"]
-	path = pystock-data
+	path = dataparser/pystock-data
 	url = https://github.com/eliangcs/pystock-data.git
