+++ dataprovider/pystock-data/.gitignore
@@ -1,0 +1,7 @@
+*.csv
+*.csv.*
+*.log
+/.scrapy/
+scrapy.cfg
+nohup.out
+
