+++ dataprovider/pystock-data/2015/index.txt
@@ -1,0 +1,202 @@
+0001_initial.tar.gz
+0002_initial.tar.gz
+0003_initial.tar.gz
+20150323.tar.gz
+20150324.tar.gz
+20150325.tar.gz
+20150326.tar.gz
+20150327.tar.gz
+20150330.tar.gz
+20150331.tar.gz
+20150401.tar.gz
+20150402.tar.gz
+20150406.tar.gz
+20150407.tar.gz
+20150408.tar.gz
+20150409.tar.gz
+20150410.tar.gz
+20150413.tar.gz
+20150414.tar.gz
+20150415.tar.gz
+20150416.tar.gz
+20150417.tar.gz
+20150420.tar.gz
+20150421.tar.gz
+20150422.tar.gz
+20150423.tar.gz
+20150424.tar.gz
+20150427.tar.gz
+20150428.tar.gz
+20150429.tar.gz
+20150430.tar.gz
+20150501.tar.gz
+20150504.tar.gz
+20150505.tar.gz
+20150506.tar.gz
+20150507.tar.gz
+20150508.tar.gz
+20150511.tar.gz
+20150512.tar.gz
+20150513.tar.gz
+20150514.tar.gz
+20150515.tar.gz
+20150518.tar.gz
+20150519.tar.gz
+20150520.tar.gz
+20150521.tar.gz
+20150522.tar.gz
+20150526.tar.gz
+20150527.tar.gz
+20150528.tar.gz
+20150529.tar.gz
+20150601.tar.gz
+20150602.tar.gz
+20150603.tar.gz
+20150604.tar.gz
+20150605.tar.gz
+20150608.tar.gz
+20150609.tar.gz
+20150610.tar.gz
+20150611.tar.gz
+20150612.tar.gz
+20150615.tar.gz
+20150616.tar.gz
+20150617.tar.gz
+20150618.tar.gz
+20150619.tar.gz
+20150622.tar.gz
+20150623.tar.gz
+20150624.tar.gz
+20150625.tar.gz
+20150626.tar.gz
+20150629.tar.gz
+20150630.tar.gz
+20150701.tar.gz
+20150702.tar.gz
+20150703.tar.gz
+20150706.tar.gz
+20150707.tar.gz
+20150708.tar.gz
+20150709.tar.gz
+20150710.tar.gz
+20150713.tar.gz
+20150714.tar.gz
+20150715.tar.gz
+20150716.tar.gz
+20150717.tar.gz
+20150720.tar.gz
+20150721.tar.gz
+20150722.tar.gz
+20150723.tar.gz
+20150724.tar.gz
+20150727.tar.gz
+20150728.tar.gz
+20150729.tar.gz
+20150730.tar.gz
+20150731.tar.gz
+20150803.tar.gz
+20150804.tar.gz
+20150805.tar.gz
+20150806.tar.gz
+20150807.tar.gz
+20150810.tar.gz
+20150811.tar.gz
+20150812.tar.gz
+20150813.tar.gz
+20150814.tar.gz
+20150817.tar.gz
+20150818.tar.gz
+20150819.tar.gz
+20150820.tar.gz
+20150821.tar.gz
+20150824.tar.gz
+20150825.tar.gz
+20150826.tar.gz
+20150827.tar.gz
+20150828.tar.gz
+20150831.tar.gz
+20150901.tar.gz
+20150902.tar.gz
+20150903.tar.gz
+20150904.tar.gz
+20150908.tar.gz
+20150909.tar.gz
+20150910.tar.gz
+20150911.tar.gz
+20150914.tar.gz
+20150915.tar.gz
+20150916.tar.gz
+20150917.tar.gz
+20150918.tar.gz
+20150921.tar.gz
+20150922.tar.gz
+20150923.tar.gz
+20150924.tar.gz
+20150925.tar.gz
+20150928.tar.gz
+20150929.tar.gz
+20150930.tar.gz
+20151001.tar.gz
+20151002.tar.gz
+20151005.tar.gz
+20151006.tar.gz
+20151007.tar.gz
+20151008.tar.gz
+20151009.tar.gz
+20151012.tar.gz
+20151013.tar.gz
+20151014.tar.gz
+20151015.tar.gz
+20151016.tar.gz
+20151019.tar.gz
+20151020.tar.gz
+20151021.tar.gz
+20151022.tar.gz
+20151023.tar.gz
+20151026.tar.gz
+20151027.tar.gz
+20151028.tar.gz
+20151029.tar.gz
+20151030.tar.gz
+20151102.tar.gz
+20151103.tar.gz
+20151104.tar.gz
+20151105.tar.gz
+20151106.tar.gz
+20151109.tar.gz
+20151110.tar.gz
+20151111.tar.gz
+20151112.tar.gz
+20151113.tar.gz
+20151116.tar.gz
+20151117.tar.gz
+20151118.tar.gz
+20151119.tar.gz
+20151120.tar.gz
+20151123.tar.gz
+20151124.tar.gz
+20151125.tar.gz
+20151127.tar.gz
+20151130.tar.gz
+20151201.tar.gz
+20151202.tar.gz
+20151203.tar.gz
+20151204.tar.gz
+20151207.tar.gz
+20151208.tar.gz
+20151209.tar.gz
+20151210.tar.gz
+20151211.tar.gz
+20151214.tar.gz
+20151215.tar.gz
+20151216.tar.gz
+20151217.tar.gz
+20151218.tar.gz
+20151221.tar.gz
+20151222.tar.gz
+20151223.tar.gz
+20151224.tar.gz
+20151228.tar.gz
+20151229.tar.gz
+20151230.tar.gz
+20151231.tar.gz
