+++ dataprovider/pystock-data/2016/index.txt
@@ -1,0 +1,253 @@
+20160104.tar.gz
+20160105.tar.gz
+20160106.tar.gz
+20160107.tar.gz
+20160108.tar.gz
+20160111.tar.gz
+20160112.tar.gz
+20160113.tar.gz
+20160114.tar.gz
+20160115.tar.gz
+20160119.tar.gz
+20160120.tar.gz
+20160121.tar.gz
+20160122.tar.gz
+20160125.tar.gz
+20160126.tar.gz
+20160127.tar.gz
+20160128.tar.gz
+20160129.tar.gz
+20160201.tar.gz
+20160202.tar.gz
+20160203.tar.gz
+20160204.tar.gz
+20160205.tar.gz
+20160208.tar.gz
+20160209.tar.gz
+20160210.tar.gz
+20160211.tar.gz
+20160212.tar.gz
+20160216.tar.gz
+20160217.tar.gz
+20160218.tar.gz
+20160219.tar.gz
+20160222.tar.gz
+20160223.tar.gz
+20160224.tar.gz
+20160225.tar.gz
+20160226.tar.gz
+20160229.tar.gz
+20160301.tar.gz
+20160302.tar.gz
+20160303.tar.gz
+20160304.tar.gz
+20160307.tar.gz
+20160308.tar.gz
+20160309.tar.gz
+20160310.tar.gz
+20160311.tar.gz
+20160314.tar.gz
+20160315.tar.gz
+20160316.tar.gz
+20160317.tar.gz
+20160318.tar.gz
+20160321.tar.gz
+20160322.tar.gz
+20160323.tar.gz
+20160324.tar.gz
+20160328.tar.gz
+20160329.tar.gz
+20160330.tar.gz
+20160331.tar.gz
+20160401.tar.gz
+20160404.tar.gz
+20160405.tar.gz
+20160406.tar.gz
+20160407.tar.gz
+20160408.tar.gz
+20160411.tar.gz
+20160412.tar.gz
+20160413.tar.gz
+20160414.tar.gz
+20160415.tar.gz
+20160418.tar.gz
+20160419.tar.gz
+20160420.tar.gz
+20160421.tar.gz
+20160422.tar.gz
+20160425.tar.gz
+20160426.tar.gz
+20160427.tar.gz
+20160428.tar.gz
+20160429.tar.gz
+20160502.tar.gz
+20160503.tar.gz
+20160504.tar.gz
+20160505.tar.gz
+20160506.tar.gz
+20160509.tar.gz
+20160510.tar.gz
+20160511.tar.gz
+20160512.tar.gz
+20160513.tar.gz
+20160516.tar.gz
+20160517.tar.gz
+20160518.tar.gz
+20160519.tar.gz
+20160520.tar.gz
+20160523.tar.gz
+20160524.tar.gz
+20160525.tar.gz
+20160526.tar.gz
+20160527.tar.gz
+20160531.tar.gz
+20160601.tar.gz
+20160602.tar.gz
+20160603.tar.gz
+20160606.tar.gz
+20160607.tar.gz
+20160608.tar.gz
+20160609.tar.gz
+20160610.tar.gz
+20160613.tar.gz
+20160614.tar.gz
+20160615.tar.gz
+20160616.tar.gz
+20160617.tar.gz
+20160620.tar.gz
+20160621.tar.gz
+20160622.tar.gz
+20160623.tar.gz
+20160624.tar.gz
+20160627.tar.gz
+20160628.tar.gz
+20160629.tar.gz
+20160630.tar.gz
+20160701.tar.gz
+20160705.tar.gz
+20160706.tar.gz
+20160707.tar.gz
+20160708.tar.gz
+20160711.tar.gz
+20160712.tar.gz
+20160713.tar.gz
+20160714.tar.gz
+20160715.tar.gz
+20160718.tar.gz
+20160719.tar.gz
+20160720.tar.gz
+20160721.tar.gz
+20160722.tar.gz
+20160725.tar.gz
+20160726.tar.gz
+20160727.tar.gz
+20160728.tar.gz
+20160729.tar.gz
+20160801.tar.gz
+20160802.tar.gz
+20160803.tar.gz
+20160804.tar.gz
+20160805.tar.gz
+20160808.tar.gz
+20160809.tar.gz
+20160810.tar.gz
+20160811.tar.gz
+20160812.tar.gz
+20160815.tar.gz
+20160816.tar.gz
+20160817.tar.gz
+20160818.tar.gz
+20160819.tar.gz
+20160822.tar.gz
+20160823.tar.gz
+20160824.tar.gz
+20160825.tar.gz
+20160826.tar.gz
+20160829.tar.gz
+20160830.tar.gz
+20160831.tar.gz
+20160901.tar.gz
+20160902.tar.gz
+20160906.tar.gz
+20160907.tar.gz
+20160908.tar.gz
+20160909.tar.gz
+20160912.tar.gz
+20160913.tar.gz
+20160914.tar.gz
+20160915.tar.gz
+20160916.tar.gz
+20160919.tar.gz
+20160920.tar.gz
+20160921.tar.gz
+20160922.tar.gz
+20160923.tar.gz
+20160926.tar.gz
+20160927.tar.gz
+20160928.tar.gz
+20160929.tar.gz
+20160930.tar.gz
+20161003.tar.gz
+20161004.tar.gz
+20161005.tar.gz
+20161006.tar.gz
+20161007.tar.gz
+20161010.tar.gz
+20161011.tar.gz
+20161012.tar.gz
+20161013.tar.gz
+20161014.tar.gz
+20161017.tar.gz
+20161018.tar.gz
+20161019.tar.gz
+20161020.tar.gz
+20161021.tar.gz
+20161024.tar.gz
+20161025.tar.gz
+20161026.tar.gz
+20161027.tar.gz
+20161028.tar.gz
+20161031.tar.gz
+20161101.tar.gz
+20161102.tar.gz
+20161103.tar.gz
+20161104.tar.gz
+20161107.tar.gz
+20161108.tar.gz
+20161109.tar.gz
+20161110.tar.gz
+20161111.tar.gz
+20161114.tar.gz
+20161115.tar.gz
+20161116.tar.gz
+20161117.tar.gz
+20161118.tar.gz
+20161121.tar.gz
+20161122.tar.gz
+20161123.tar.gz
+20161125.tar.gz
+20161128.tar.gz
+20161129.tar.gz
+20161130.tar.gz
+20161201.tar.gz
+20161202.tar.gz
+20161205.tar.gz
+20161206.tar.gz
+20161207.tar.gz
+20161208.tar.gz
+20161209.tar.gz
+20161212.tar.gz
+20161213.tar.gz
+20161214.tar.gz
+20161215.tar.gz
+20161216.tar.gz
+20161219.tar.gz
+20161220.tar.gz
+20161221.tar.gz
+20161222.tar.gz
+20161223.tar.gz
+20161226.tar.gz
+20161227.tar.gz
+20161228.tar.gz
+20161229.tar.gz
+20161230.tar.gz
