+++ dataprovider/pystock-data/2017/index.html
@@ -1,0 +1,65 @@
+<html><body>
+<a href="20170102.tar.gz">20170102.tar.gz</a><br>
+<a href="20170103.tar.gz">20170103.tar.gz</a><br>
+<a href="20170104.tar.gz">20170104.tar.gz</a><br>
+<a href="20170105.tar.gz">20170105.tar.gz</a><br>
+<a href="20170106.tar.gz">20170106.tar.gz</a><br>
+<a href="20170109.tar.gz">20170109.tar.gz</a><br>
+<a href="20170110.tar.gz">20170110.tar.gz</a><br>
+<a href="20170111.tar.gz">20170111.tar.gz</a><br>
+<a href="20170112.tar.gz">20170112.tar.gz</a><br>
+<a href="20170113.tar.gz">20170113.tar.gz</a><br>
+<a href="20170117.tar.gz">20170117.tar.gz</a><br>
+<a href="20170118.tar.gz">20170118.tar.gz</a><br>
+<a href="20170119.tar.gz">20170119.tar.gz</a><br>
+<a href="20170120.tar.gz">20170120.tar.gz</a><br>
+<a href="20170123.tar.gz">20170123.tar.gz</a><br>
+<a href="20170124.tar.gz">20170124.tar.gz</a><br>
+<a href="20170125.tar.gz">20170125.tar.gz</a><br>
+<a href="20170126.tar.gz">20170126.tar.gz</a><br>
+<a href="20170127.tar.gz">20170127.tar.gz</a><br>
+<a href="20170130.tar.gz">20170130.tar.gz</a><br>
+<a href="20170131.tar.gz">20170131.tar.gz</a><br>
+<a href="20170201.tar.gz">20170201.tar.gz</a><br>
+<a href="20170202.tar.gz">20170202.tar.gz</a><br>
+<a href="20170203.tar.gz">20170203.tar.gz</a><br>
+<a href="20170206.tar.gz">20170206.tar.gz</a><br>
+<a href="20170207.tar.gz">20170207.tar.gz</a><br>
+<a href="20170208.tar.gz">20170208.tar.gz</a><br>
+<a href="20170209.tar.gz">20170209.tar.gz</a><br>
+<a href="20170210.tar.gz">20170210.tar.gz</a><br>
+<a href="20170213.tar.gz">20170213.tar.gz</a><br>
+<a href="20170214.tar.gz">20170214.tar.gz</a><br>
+<a href="20170215.tar.gz">20170215.tar.gz</a><br>
+<a href="20170216.tar.gz">20170216.tar.gz</a><br>
+<a href="20170217.tar.gz">20170217.tar.gz</a><br>
+<a href="20170221.tar.gz">20170221.tar.gz</a><br>
+<a href="20170222.tar.gz">20170222.tar.gz</a><br>
+<a href="20170223.tar.gz">20170223.tar.gz</a><br>
+<a href="20170224.tar.gz">20170224.tar.gz</a><br>
+<a href="20170227.tar.gz">20170227.tar.gz</a><br>
+<a href="20170228.tar.gz">20170228.tar.gz</a><br>
+<a href="20170301.tar.gz">20170301.tar.gz</a><br>
+<a href="20170302.tar.gz">20170302.tar.gz</a><br>
+<a href="20170303.tar.gz">20170303.tar.gz</a><br>
+<a href="20170306.tar.gz">20170306.tar.gz</a><br>
+<a href="20170307.tar.gz">20170307.tar.gz</a><br>
+<a href="20170308.tar.gz">20170308.tar.gz</a><br>
+<a href="20170309.tar.gz">20170309.tar.gz</a><br>
+<a href="20170310.tar.gz">20170310.tar.gz</a><br>
+<a href="20170313.tar.gz">20170313.tar.gz</a><br>
+<a href="20170314.tar.gz">20170314.tar.gz</a><br>
+<a href="20170315.tar.gz">20170315.tar.gz</a><br>
+<a href="20170316.tar.gz">20170316.tar.gz</a><br>
+<a href="20170317.tar.gz">20170317.tar.gz</a><br>
+<a href="20170320.tar.gz">20170320.tar.gz</a><br>
+<a href="20170321.tar.gz">20170321.tar.gz</a><br>
+<a href="20170322.tar.gz">20170322.tar.gz</a><br>
+<a href="20170323.tar.gz">20170323.tar.gz</a><br>
+<a href="20170324.tar.gz">20170324.tar.gz</a><br>
+<a href="20170327.tar.gz">20170327.tar.gz</a><br>
+<a href="20170328.tar.gz">20170328.tar.gz</a><br>
+<a href="20170329.tar.gz">20170329.tar.gz</a><br>
+<a href="20170330.tar.gz">20170330.tar.gz</a><br>
+<a href="20170331.tar.gz">20170331.tar.gz</a><br>
+</body></html>
