+++ dataprovider/pystock-data/CNAME
@@ -1,0 +1,1 @@
+data.pystock.com
