+++ dataprovider/pystock-data/LICENSE
@@ -1,0 +1,8 @@
+pystock-data (c) by Chang-Hung Liang
+
+pystock-data is licensed under a
+Creative Commons Attribution-ShareAlike 4.0 International License.
+
+You should have received a copy of the license along with this
+work. If not, see <http://creativecommons.org/licenses/by-sa/4.0/>. 
+
