+++ dataprovider/pystock-data/index.html
@@ -1,0 +1,6 @@
+<html><body>
+<a href="2015">2015</a><br>
+<a href="2016">2016</a><br>
+<a href="2017">2017</a><br>
+<a href="README.md">README.md</a><br>
+</body></html>
