+++ dataprovider/pystock-data/index.txt
@@ -1,0 +1,4 @@
+2015
+2016
+2017
+README.md
