+++ dataprovider/src/main/java/qwde/dataprovider/pystock/PystockSelection.java
@@ -1,0 +1,10 @@
+package qwde.dataprovider.pystock;
+
+import java.io.IOException;
+import java.time.LocalDate;
+
+public class PystockSelection {
+    public PystockSelection(String ticker, LocalDate fromDate, LocalDate toDate) throws IOException {
+        PystockStockPriceReader pystockStockPriceReader = PystockStockPriceReader.FromDate(fromDate);
+    }
+}
