+++ dataprovider/src/main/java/qwde/dataprovider/util/StockPriceReader.java
@@ -1,0 +1,9 @@
+package qwde.dataprovider.util; 
+
+import java.util.Collection;
+
+import qwde.dataprovider.models.StockPrice;
+
+public interface StockPriceReader {
+  Collection<StockPrice> read();
+}
