--- qwde.pystock/src/main/java/qwdepystock/pystock/PystockSelection.java
@@ -1,10 +1,0 @@
-package qwdepystock.pystock;
-
-import java.io.IOException;
-import java.time.LocalDate;
-
-public class PystockSelection {
-    public PystockSelection(String ticker, LocalDate fromDate, LocalDate toDate) throws IOException {
-        PystockStockPriceReader pystockStockPriceReader = PystockStockPriceReader.FromDate(fromDate);
-    }
-}
