--- qwde.pystock/src/main/java/qwdepystock/util/StockPriceReader.java
@@ -1,9 +1,0 @@
-package qwdepystock.util;
-
-import java.util.Collection;
-
-import qwdepystock.models.StockPrice;
-
-public interface StockPriceReader {
-  Collection<StockPrice> read();
-}
