+++ settings.gradle.kts
@@ -1,5 +1,5 @@
 rootProject.name = "stockerface"
 
 include("qwde")
-include("qwde.pystock")
+include("dataprovider")
 include("main")
